Задание 3.1 выполнено...
