+++ templates/layout.php
@@ -1,8 +1,3 @@
-<?php
-$is_auth = rand(0, 1);
-
-$user_name = 'Ihor'; // укажите здесь ваше имя
-?>
 <!DOCTYPE html>
 <html lang="ru">
 <head>
@@ -56,9 +51,6 @@ $user_name = 'Ihor'; // укажите здесь ваше имя
     <nav class="nav">
         <ul class="nav__list container">
             <!--заполните этот список из массива категорий-->
-            <?php
-                $categories = ['Доски и лыжи', 'Крепления', 'Ботинки', 'Одежда', 'Инструменты', 'Разное'];
-            ?>
             <?php foreach($categories as $item) : ?>
             <li class="nav__item">
                 <a href="pages/#"><?= $item ?></a>
